pyproject.toml
[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[project]
name = "redcap_toolbox"
requires-python = ">= 3.10"
version = "2025.09.10"
authors = [
    { name="Nate Vack", email="njvack@wisc.edu"},
    { name="Stuti Shrivastava", email="sshrivastav6@wisc.edu" },
    { name="Nicholas Vanhaute", email="nvanhaute@wisc.edu" },
]
maintainers = [
    { name="Nate Vack", email="njvack@wisc.edu"},
    { name="Stuti Shrivastava", email="sshrivastav6@wisc.edu" },
]
description = "A collection of tools to interact with the REDCap API"
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "docopt-ng>=0.9.0",
    "polars>=1.36.1",
    "pycap",
]


[project.urls]
Homepage = "https://github.com/uwmadison-chm/redcap-toolbox"
Issues = "https://github.com/uwmadison-chm/redcap-toolbox/issues"

[project.scripts]
download_redcap = "redcap_toolbox.download_redcap:main"
split_redcap_data = "redcap_toolbox.split_redcap_data:main"
update_redcap_diff = "redcap_toolbox.update_redcap_diff:main"
download_redcap_report = "redcap_toolbox.download_redcap_report:main"
download_redcap_incremental = "redcap_toolbox.download_redcap_incremental:main"

[dependency-groups]
dev = [
    "pluggy>=1.6.0",
    "pytest>=8.4.1",
    "ruff>=0.12.2",
    "ty>=0.0.2",
]

[tool.uv.sources]
pycap = { url = "https://github.com/uwmadison-chm/PyCap/archive/refs/tags/2.7.1-chm.zip" }

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
